requirements.txt
asgiref==3.3.4
astroid==2.6.2
attrs==21.2.0
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.4
coverage==5.5
cryptography==3.4.7
cssutils==2.3.0
Django==3.2.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.7.0
django-inlinecss==0.3.0
django-treebeard==4.5.1
djangorestframework==3.12.4
flake8==3.9.2
future==0.18.2
google-api-core==1.31.1
google-auth==1.34.0
google-cloud==0.34.0
google-cloud-core==1.7.2
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
idna==3.2
iniconfig==1.1.1
isort==5.9.2
jsonfield==3.1.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
numpy==1.21.1
packaging==21.0
Pillow==8.3.1
pluggy==0.13.1
protobuf==3.17.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pylint==2.9.3
PyMySQL==1.0.2
pynliner==0.8.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
python-dotenv==0.18.0
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
wrapt==1.12.1
psycopg2==2.9.1
gunicorn==20.1.0
dj-static==0.0.6
whitenoise==5.3.0